requirements.txt
langchain
openai
chromadb
streamlit
PyMuPDF
python-dotenv
tiktoken
